+++ src/main/scala/mesosphere/marathon/MarathonModule.scala
@@ -28,7 +28,6 @@ import scala.util.control.NonFatal
 object ModuleNames {
   final val HOST_PORT = "HOST_PORT"
 
-  final val SERVER_SET_PATH = "SERVER_SET_PATH"
   final val HISTORY_ACTOR_PROPS = "HISTORY_ACTOR_PROPS"
 
   final val MESOS_HEARTBEAT_ACTOR = "MesosHeartbeatActor"
@@ -54,10 +53,6 @@ class MarathonModule(conf: MarathonConf, http: HttpConf, actorSystem: ActorSyste
     bind(classOf[SchedulerDriverFactory]).to(classOf[MesosSchedulerDriverFactory]).in(Scopes.SINGLETON)
     bind(classOf[MarathonSchedulerService]).in(Scopes.SINGLETON)
     bind(classOf[DeploymentService]).to(classOf[MarathonSchedulerService])
-
-    bind(classOf[String])
-      .annotatedWith(Names.named(ModuleNames.SERVER_SET_PATH))
-      .toInstance(conf.zooKeeperServerSetPath)
   }
 
   @Named(ModuleNames.MESOS_HEARTBEAT_ACTOR)